+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `artisan-commands` will be documented in this file
 
+## 1.2.0 - 2020-02-18
+- Make directory feature available. <br>
+`php artisan delete:dir`
+
 ## 1.1.2 - 2020-02-17
 - Update README.md with correct links